Fixing add_all to be robust to git-ignored directories, git_fs(_shim) cleanup (#1363) (#1457)
* Move subprocess calls from git_fs to git_fs_shim

* Include check for untracked in `add_all()`
