+++ changelog/releases/v1.3.3/1416-fullscreen-note-editor.yaml
